// Downloads/TeachingAuraWebs-master/TeachingAuraWebs-master/src/client/models/BatchDetails.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type BatchDetails = {
    id: string;
    name: string;
    startDate?: number;
    courseId: string;
    courseName: string;
    instituteId: string;
    imageUrl?: string;
    numberOfTeachers?: number;
    numberOfStudents?: number;
    active?: boolean;
};