// Downloads/TeachingAuraWebs-master/TeachingAuraWebs-master/src/client/models/ScheduleDetails.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ScheduleDetails = {
    id: string;
    instituteId: string;
    batchId: string;
    courseId: string;
    teacherId: string;
    topicName: string;
    startTime?: number;
    endTime?: number;
    joinUrl?: string;
    meetingId?: string;
};